feat: introduce block-level markdown processing with support for tables, lists, images, and alignment
